requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
coverage==5.5
Django==3.2.6
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
idna==3.2
lxml==4.6.3
Pillow==8.3.1
psycopg2==2.9.1
PyJWT==2.1.0
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
transliterate==1.10.2
urllib3==1.26.6
